refacto: split tests into test_meals_get.py and test_meals_post.py
